refactor List summary data tests
